// packages/shared/src/types.ts
export enum CardKind {
  MoveTo = 'MoveTo',
  MoveToNearestRailroad = 'MoveToNearestRailroad',
  MoveToNearestUtility = 'MoveToNearestUtility',
  Pay = 'Pay',
  Receive = 'Receive',
  PayEachPlayer = 'PayEachPlayer',
  GoToJail = 'GoToJail',
  GetOutOfJail = 'GetOutOfJail',
  Repairs = 'Repairs'
}

export type Card =
  | { kind: CardKind.MoveTo; arg: number; text: string }
  | { kind: CardKind.MoveToNearestRailroad; text: string }
  | { kind: CardKind.MoveToNearestUtility; text: string }
  | { kind: CardKind.Pay; arg: number; text: string }
  | { kind: CardKind.Receive; arg: number; text: string }
  | { kind: CardKind.PayEachPlayer; arg: number; text: string }
  | { kind: CardKind.GoToJail; text: string }
  | { kind: CardKind.GetOutOfJail; text: string }
  | { kind: CardKind.Repairs; arg: number; arg2: number; text: string }

export type CardDeck = { seed: string; chance: Card[]; chest: Card[]; chanceIndex: number; chestIndex: number }